Hotfix for Buybinds causing decoy priority on 4.
Recent behavior change added the ability to quick switch to decoy grenades with the use of 4 when you have one equipped, much like the other grenades; but would still function like the old bind when you did not.

However, this feature would not continue to work when you pressed the buy binds key. It would simply revert to the original behavior and ignore decoys.